refactored CutomLabelGenerator
