+++ dl4j-cv-labs/src/main/java/global/skymind/solution/segmentation/car/CarDataSetIterator.java
@@ -3,13 +3,6 @@ package global.skymind.solution.segmentation.car;
 import global.skymind.Helper;
 import org.datavec.image.transform.ImageTransform;
 import org.deeplearning4j.datasets.datavec.RecordReaderDataSetIterator;
-import net.lingala.zip4j.core.ZipFile;
-import net.lingala.zip4j.exception.ZipException;
-import org.apache.http.HttpEntity;
-import org.apache.http.client.methods.CloseableHttpResponse;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClientBuilder;
 import org.datavec.api.io.filters.BalancedPathFilter;
 import org.datavec.api.split.FileSplit;
 import org.datavec.api.split.InputSplit;
@@ -17,11 +10,13 @@ import org.datavec.image.loader.NativeImageLoader;
 import org.datavec.image.recordreader.ImageRecordReader;
 import org.nd4j.linalg.dataset.api.preprocessor.DataNormalization;
 import org.nd4j.linalg.dataset.api.preprocessor.ImagePreProcessingScaler;
+import global.skymind.solution.segmentation.imageUtils.CustomLabelGenerator;
 
 import java.io.File;
-import java.io.FileOutputStream;
 import java.io.IOException;
 import java.nio.file.Paths;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Random;
 
 public class CarDataSetIterator {
@@ -31,10 +26,17 @@ public class CarDataSetIterator {
     private static final long seed = 12345;
     private static final Random random = new Random(seed);
     private static String inputDir;
-    private static String downloadLink;
-    private static CustomLabelGenerator labelMaker = new CustomLabelGenerator(height, width, channels);
+//    private static String labelPath;
+//    private static String downloadLink;
+
+
     private static InputSplit trainData,valData;
     private static int batchSize;
+    private static List<org.nd4j.linalg.primitives.Pair<String, String>> replacement = Arrays.asList(
+            new org.nd4j.linalg.primitives.Pair<>("inputs","masks"),
+            new org.nd4j.linalg.primitives.Pair<>(".jpg","_mask.gif")
+    );
+    private static CustomLabelGenerator labelMaker = new CustomLabelGenerator(height, width, channels,replacement);
 
     //scale input to 0 - 1
     private static DataNormalization scaler = new ImagePreProcessingScaler(0, 1);
@@ -45,6 +47,7 @@ public class CarDataSetIterator {
 
     //This method instantiates an ImageRecordReader and subsequently a RecordReaderDataSetIterator based on it
     private static RecordReaderDataSetIterator makeIterator(InputSplit split, boolean training) throws IOException {
+
         ImageRecordReader recordReader = new ImageRecordReader(height,width,channels,labelMaker);
         if (training && transform != null){
             recordReader.initialize(split,transform);
@@ -70,9 +73,7 @@ public class CarDataSetIterator {
         setup(batchSizeArg,trainPerc);
     }
 
-    //This method does the following:
-    // 1. Download and unzip dataset if it hasn't been downloaded
-    // 2. Split dataset into training set and validation set
+
     public static void setup(int batchSizeArg, double trainPerc) throws IOException {
 
         batchSize = batchSizeArg;
@@ -82,10 +83,15 @@ public class CarDataSetIterator {
                 Helper.getPropValues("dl4j_home.data")
         ).toString();
 
+
         File imagesPath = new File(Paths.get(inputDir, "carvana-image-masking-challenge","train","inputs").toString());
         FileSplit imageFileSplit = new FileSplit(imagesPath, NativeImageLoader.ALLOWED_FORMATS, random);
+
         BalancedPathFilter imageSplitPathFilter = new BalancedPathFilter(random, NativeImageLoader.ALLOWED_FORMATS, labelMaker);
         InputSplit[] imagesSplits = imageFileSplit.sample(imageSplitPathFilter, trainPerc, 1-trainPerc);
+
+//        labelPath = imageFileSplit.toString().replace("inputs", "masks").replace(".jpg","_mask.gif");
+//        labelMaker = new CustomLabelGenerator(height, width, channels, labelPath);
 
         trainData = imagesSplits[0];
         valData = imagesSplits[1];
